third_party/WebKit/LayoutTests/inspector-protocol/runtime/runtime-evaluate-async-expected.txt

Tests that Runtime.evaluate works with awaitPromise flag.

Running test: testResolvedPromise
{
    result : {
        description : 239
        type : number
        value : 239
    }
}

Running test: testRejectedPromise
{
    exceptionDetails : {
        columnNumber : 0
        exception : {
            description : 239
            objectId : 0
            type : number
            value : 239
        }
        exceptionId : 0
        lineNumber : 0
        scriptId : 
        text : Uncaught (in promise)
    }
    result : {
        description : 239
        type : number
        value : 239
    }
}

Running test: testPrimitiveValueInsteadOfPromise
{
    result : {
        type : boolean
        value : true
    }
}

Running test: testObjectInsteadOfPromise
{
    result : {
        type : object
        value : {
        }
    }
}

Running test: testPendingPromise
{
    result : {
        type : object
        value : {
            a : 239
        }
    }
}

Running test: testExceptionInEvaluate
{
    exceptionDetails : {
        columnNumber : 0
        exception : {
            description : 239
            objectId : 0
            type : number
            value : 239
        }
        exceptionId : 0
        lineNumber : 0
        scriptId : 
        text : Uncaught
    }
    result : {
        description : 239
        type : number
        value : 239
    }
}
